requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
pytesseract>=0.3.10
SpeechRecognition>=3.10.0
pyaudio>=0.2.14
openai>=1.0.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
